organize things better
