+++ optout/src/main/java/gov/cms/ab2d/optout/gateway/S3GatewayImpl.java
@@ -1,4 +1,4 @@
-package gov.cms.ab2d.optout;
+package gov.cms.ab2d.optout.gateway;
 
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Value;
