+++ optout/src/test/java/gov/cms/ab2d/optout/OptOutProcessorIntegrationTest.java
@@ -1,10 +1,10 @@
 package gov.cms.ab2d.optout;
 
 import gov.cms.ab2d.common.repository.ConsentRepository;
+import gov.cms.ab2d.optout.gateway.S3Gateway;
 import lombok.extern.slf4j.Slf4j;
 import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mockito;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.boot.test.context.SpringBootTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
